+++ src/components/ViewCodeFile.vue
@@ -10,8 +10,7 @@
     >
       <i class="icon-download-cloud icon--scaling"></i>
     </a>
-    <pre v-if="!file.meta.highlight || file.meta.highlight === 'auto'"><code ref="code">{{ file.data }}</code></pre>
-    <pre v-else><code :class=file.meta.highlight ref="code">{{ file.data }}</code></pre>
+    <pre><code :class=file.meta.highlight ref="code">{{ file.data }}</code></pre>
   </div>
 </template>
 
@@ -22,12 +21,8 @@
  export default {
    props: ['file'],
    mounted() {
-     if(this.file.meta.highlight != 'plain') {
-       hljs.highlightBlock(this.$refs.code);
-     } else {
-       $(this.$refs.code).addClass('hljs');
-     }
-
+     hljs.highlightBlock(this.$refs.code);
+     $(this.$refs.code).addClass('hljs');
      helpers.splitWithLineNumbers(this.$refs.code);
 
      $('.hljs-link').each((i, e) => {
@@ -36,7 +31,7 @@
 
      const bgcolor = $(this.$refs.code).css('background-color');
      $(this.$refs.code).find('tr').css('background', bgcolor);
-     // $(this.$refs.code).css('background', 'transparent');
+     $(this.$refs.code).css('background', 'transparent');
 
      helpers.registerClickHandlers($(this.$refs.code));